requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py38h06a4308_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py38h295c915_0
decorator=5.1.0=pypi_0
elasticdeform=0.4.9=pypi_0
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnureadline=8.0.0=pypi_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
intel-openmp=2021.3.0=h06a4308_3350
ipdb=0.13.9=pypi_0
ipykernel=6.2.0=py38h06a4308_1
ipython=7.27.0=py38hb070fc8_0
jedi=0.18.0=py38h06a4308_1
jpeg=9b=h024ee3a_2
jupyter_client=7.0.1=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.4.2=py38h06a4308_0
matplotlib-base=3.4.2=py38hab158f2_0
matplotlib-inline=0.1.3=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.2=py38h8c16a72_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py38h2c7a002_0
pip=21.0.1=py38h06a4308_0
prompt-toolkit=3.0.20=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pygments=2.10.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.2.1=py38h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scipy=1.7.1=py38h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tornado=6.1=py38h27cfd23_0
tqdm=4.62.1=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
typing_extensions=3.10.0.0=pyhca03da5_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
